Refactor layoutAttributesForElements to do proper dequeing
On macOS, the collection view is the document view of a scroll view, to get proper dequeuing we need to resolve the scroll views rectangle instead of the rectangle that is passed to the collection view layout.

This way we make sure that we never allocate more items than necessary.