Fix crash when reloading collection view
Remove isUpdating optimization which led to crashes when reloading collection views. The occured because the optimization outed out from returning anything other than `nil` when and update is happening. Internally this caused the collection view to throw an exception when updating a collection view with mutliple inserts and deletions.

Similar to this issue here: https://github.com/ra1028/DifferenceKit/issues/13