Implemented check to ensure that the index is not out of bounds when obtaining the chached attributes.
There is a case when the transtion and cache may be cleared at the same time causing a possible index out of bounds. At the same time it has been observed that the indexPath may also be invalidated by the OS so an intial check for the cachedAttributes needs to be made before we check that the section is greater than the cachedAttributes.