[WIP] - Adding dynamic cell height configuration
Noticed a bug in the vertical layout, if the previous collumns last item has a larger height than the last item in the next column the collection view, then the item will be cut off. (scroll view bound to the last item in the last collumn)

Likewise similar display issues are seen with the same configuration when using a horizontal layout.