Add padding to BinarySearch.findElements
When using multiple items per row with dynamic height, the binary search can cause UI glitches where certain items simply disappear after a certain content offset. To fix this issue, the `findElements` method can now pad the results that it finds. The itemsPerRow is used for padding. What padding means in this context is the number of extra items that should be allowed to exist even if they don't meet the search criteria. That way we can capture all views that are on the same row even if they have different y,x coordinates. In a real-world scenario, the number of padded items should never exceed four or five, tops.